+++ Assets/Scripts/Gamecore/Card/AttributeCard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 759b1d826adc58e41aec9f4fbf5dcfa6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
