+++ Assets/Scripts/Gamecore/Card/CardBase.cs.meta
