+++ Assets/Scripts/Gamecore/Card/MonsterCard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a30045ee4bbc864c95aa5ce8cb84644
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
